add more tests
These tests are an adaptation from the pre-existing tcsh tests.
The tests have been migrated to bash and reformulated to be more
appropriate for continuous integration testing. Where the command
is expected to rais an error, it is prepended with ! to pass and
stderr is parsed to confirm that the failure was due to the appropriate
error during execution.

The tests should be run with -e as no error should occur during
running of the tests.